+++ src/bika/lims/setuphandlers.py
@@ -510,6 +510,10 @@ def add_dexterity_setup_items(portal):
         ("dynamic_analysisspecs",  # ID
          "Dynamic Analysis Specifications",  # Title
          "DynamicAnalysisSpecs"),  # FTI
+
+        ("interpretation_templates",
+         "Interpretation Templates",
+         "InterpretationTemplates")
     ]
 
     for id, title, fti in items: